+++ src/hiero_sdk_python/__init__.py
@@ -5,6 +5,7 @@ from .client.network import Network
 # Account
 from .account.account_id import AccountId
 from .account.account_create_transaction import AccountCreateTransaction
+from .account.account_update_transaction import AccountUpdateTransaction
 from .account.account_info import AccountInfo
 
 # Crypto
@@ -96,6 +97,7 @@ from .file.file_delete_transaction import FileDeleteTransaction
 from .contract.contract_create_transaction import ContractCreateTransaction
 from .contract.contract_call_query import ContractCallQuery
 from .contract.contract_info_query import ContractInfoQuery
+from .contract.contract_bytecode_query import ContractBytecodeQuery
 from .contract.contract_function_parameters import ContractFunctionParameters
 from .contract.contract_function_result import ContractFunctionResult
 from .contract.contract_info import ContractInfo
@@ -108,6 +110,7 @@ __all__ = [
     # Account
     "AccountId",
     "AccountCreateTransaction",
+    "AccountUpdateTransaction",
     "AccountInfo",
 
     # Crypto
@@ -195,6 +198,7 @@ __all__ = [
     "ContractCreateTransaction",
     "ContractCallQuery",
     "ContractInfoQuery",
+    "ContractBytecodeQuery",
     "ContractFunctionParameters",
     "ContractFunctionResult",
     "ContractInfo"