+++ CHANGELOG.md
@@ -24,11 +24,17 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - PendingAirdropRecord Class
 - TokenCancelAirdropTransaction Class
 - Integrated pylint.yml and test.yml into a single workflow file ci.yml
+- AccountUpdateTransaction class
+- ContractBytecodeQuery class
+- SimpleStorage.bin-runtime
+- Support for both .bin and .bin-runtime contract bytecode extensions in contract_utils.py
 
 ### Fixed
 
 - missing ECDSA support in query.py and contract_create_transaction.py (was only creating ED25519 keys)
 - Applied linting and code formatting across the consensus module
+- fixed pip install hiero_sdk_python -> pip install hiero-sdk-python in README.md
+
 
 ### Breaking API changes